pyproject.toml
[project]
name = "pdf-absct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "click>=8.3.3",
    "langgraph>=1.1.9",
    "openai>=2.32.0",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]
[project.scripts]
extract = "source.main:cli"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]
[tool.pytest.ini_options]
pythonpath = ["."]
